Update stochastic_breakout.py
What's new:
- When you destroy all the bricks, it switches to the next lvl within 1 sec. If there is no next lvl, the game close.
Problem to solve: When switching to next lvl, we want the ball to return to the paddle in order to start the new lvl as we started the first (with the ball on the paddle, still). I found how to put the ball on the paddle (with Ball(arena, paddle, bricks)) but the previous ball won't disappear automatically. I can't find how to make it so.

- I found a really simple menu (cf. line 274 to 336 and the new file ezmenu.py) and I think it could be of some use for us. It is not implemented yet.