Update stochastic_breakout.py
Minor change: the aspect of the ball (now gray)

After destroying the last brick on the screen, there is a little delay (customable) and then the next lvl is charged (correction of all previous issues). If there is no further lvl, the game return to the main menu.

Adding of a random component on the rebound when destroying a brick (cf. class Ball; def move; line 225 and further).

Remaining problems:
- when using EyeTribe, the game is greatly slowed.
- the gaze coordinates don't get written in the 'data_eye_tracking.dat' file (cf. line 534).